# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "integr8scode",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/replay/sessions": {
      "post": {
        "tags": [
          "Event Replay"
        ],
        "summary": "Create Replay Session",
        "operationId": "create_replay_session_api_v1_replay_sessions_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Event Replay"
        ],
        "summary": "List Replay Sessions",
        "operationId": "list_replay_sessions_api_v1_replay_sessions_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReplayStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionSummary"
                  },
                  "title": "Response List Replay Sessions Api V1 Replay Sessions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replay/sessions/{session_id}/start": {
      "post": {
        "tags": [
          "Event Replay"
        ],
        "summary": "Start Replay Session",
        "operationId": "start_replay_session_api_v1_replay_sessions__session_id__start_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replay/sessions/{session_id}/pause": {
      "post": {
        "tags": [
          "Event Replay"
        ],
        "summary": "Pause Replay Session",
        "operationId": "pause_replay_session_api_v1_replay_sessions__session_id__pause_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replay/sessions/{session_id}/resume": {
      "post": {
        "tags": [
          "Event Replay"
        ],
        "summary": "Resume Replay Session",
        "operationId": "resume_replay_session_api_v1_replay_sessions__session_id__resume_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replay/sessions/{session_id}/cancel": {
      "post": {
        "tags": [
          "Event Replay"
        ],
        "summary": "Cancel Replay Session",
        "operationId": "cancel_replay_session_api_v1_replay_sessions__session_id__cancel_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replay/sessions/{session_id}": {
      "get": {
        "tags": [
          "Event Replay"
        ],
        "summary": "Get Replay Session",
        "operationId": "get_replay_session_api_v1_replay_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplaySession"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/replay/cleanup": {
      "post": {
        "tags": [
          "Event Replay"
        ],
        "summary": "Cleanup Old Sessions",
        "operationId": "cleanup_old_sessions_api_v1_replay_cleanup_post",
        "parameters": [
          {
            "name": "older_than_hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 24,
              "title": "Older Than Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CleanupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CleanupResponse": {
        "properties": {
          "removed_sessions": {
            "type": "integer",
            "title": "Removed Sessions"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "removed_sessions",
          "message"
        ],
        "title": "CleanupResponse",
        "description": "Response schema for cleanup operations"
      },
...
      "EventType": {
        "type": "string",
        "enum": [
          "execution_requested",
          "execution_accepted",
          "execution_queued",
          "execution_started",
          "execution_running",
          "execution_completed",
          "execution_failed",
          "execution_timeout",
          "execution_cancelled",
          "pod_created",
          "pod_scheduled",
          "pod_running",
          "pod_succeeded",
          "pod_failed",
          "pod_terminated",
          "pod_deleted",
          "user_registered",
          "user_login",
          "user_logged_in",
          "user_logged_out",
          "user_updated",
          "user_deleted",
          "user_settings_updated",
          "user_theme_changed",
          "user_notification_settings_updated",
          "user_editor_settings_updated",
          "notification_created",
          "notification_sent",
          "notification_delivered",
          "notification_failed",
          "notification_read",
          "notification_clicked",
          "notification_preferences_updated",
          "script_saved",
          "script_deleted",
          "script_shared",
          "security_violation",
          "rate_limit_exceeded",
          "auth_failed",
          "resource_limit_exceeded",
          "quota_exceeded",
          "system_error",
          "service_unhealthy",
          "service_recovered",
          "result_stored",
          "result_failed",
          "saga_started",
          "saga_completed",
          "saga_failed",
          "saga_cancelled",
          "saga_compensating",
          "saga_compensated",
          "create_pod_command",
          "delete_pod_command",
          "allocate_resources_command",
          "release_resources_command"
        ],
        "title": "EventType",
        "description": "Event types used throughout the system."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReplayConfigSchema": {
        "properties": {
          "replay_type": {
            "$ref": "#/components/schemas/ReplayType"
          },
          "target": {
            "$ref": "#/components/schemas/ReplayTarget",
            "default": "kafka"
          },
          "filter": {
            "$ref": "#/components/schemas/ReplayFilterSchema"
          },
          "speed_multiplier": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.1,
            "title": "Speed Multiplier",
            "default": 1.0
          },
          "preserve_timestamps": {
            "type": "boolean",
            "title": "Preserve Timestamps",
            "default": false
          },
          "batch_size": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Batch Size",
            "default": 100
          },
          "max_events": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Events"
          },
          "target_topics": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Topics"
          },
          "target_file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target File Path"
          },
          "skip_errors": {
            "type": "boolean",
            "title": "Skip Errors",
            "default": true
          },
          "retry_failed": {
            "type": "boolean",
            "title": "Retry Failed",
            "default": false
          },
          "retry_attempts": {
            "type": "integer",
            "title": "Retry Attempts",
            "default": 3
          },
          "enable_progress_tracking": {
            "type": "boolean",
            "title": "Enable Progress Tracking",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "replay_type"
        ],
        "title": "ReplayConfigSchema"
      },
...
      "ReplayFilter": {
        "properties": {
          "execution_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Id"
          },
          "event_types": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EventType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Types"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "service_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Name"
          },
          "custom_query": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Query"
          },
          "exclude_event_types": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EventType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclude Event Types"
          }
        },
        "type": "object",
        "title": "ReplayFilter"
      },
      "ReplayFilterSchema": {
        "properties": {
          "execution_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Id"
          },
          "event_types": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EventType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Types"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          },
          "service_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Name"
          },
          "custom_query": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Custom Query"
          },
          "exclude_event_types": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/EventType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclude Event Types"
          }
        },
        "type": "object",
        "title": "ReplayFilterSchema"
      },
      "ReplayRequest": {
        "properties": {
          "replay_type": {
            "$ref": "#/components/schemas/ReplayType"
          },
          "target": {
            "$ref": "#/components/schemas/ReplayTarget",
            "default": "kafka"
          },
          "filter": {
            "$ref": "#/components/schemas/ReplayFilter"
          },
          "speed_multiplier": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.1,
            "title": "Speed Multiplier",
            "default": 1.0
          },
          "preserve_timestamps": {
            "type": "boolean",
            "title": "Preserve Timestamps",
            "default": false
          },
          "batch_size": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Batch Size",
            "default": 100
          },
          "max_events": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Events"
          },
          "skip_errors": {
            "type": "boolean",
            "title": "Skip Errors",
            "default": true
          },
          "target_file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target File Path"
          },
          "target_topics": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Topics"
          },
          "retry_failed": {
            "type": "boolean",
            "title": "Retry Failed",
            "default": false
          },
          "retry_attempts": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Retry Attempts",
            "default": 3
          },
          "enable_progress_tracking": {
            "type": "boolean",
            "title": "Enable Progress Tracking",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "replay_type"
        ],
        "title": "ReplayRequest",
        "description": "Request schema for creating replay sessions"
      },
      "ReplayResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "$ref": "#/components/schemas/ReplayStatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status",
          "message"
        ],
        "title": "ReplayResponse",
        "description": "Response schema for replay operations"
      },
      "ReplaySession": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "config": {
            "$ref": "#/components/schemas/ReplayConfigSchema"
          },
          "status": {
            "$ref": "#/components/schemas/ReplayStatus",
            "default": "created"
          },
          "total_events": {
            "type": "integer",
            "title": "Total Events",
            "default": 0
          },
          "replayed_events": {
            "type": "integer",
            "title": "Replayed Events",
            "default": 0
          },
          "failed_events": {
            "type": "integer",
            "title": "Failed Events",
            "default": 0
          },
          "skipped_events": {
            "type": "integer",
            "title": "Skipped Events",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "last_event_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event At"
          },
          "errors": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "config"
        ],
        "title": "ReplaySession"
      },
      "ReplayStatus": {
        "type": "string",
        "enum": [
          "scheduled",
          "created",
          "running",
          "paused",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "ReplayStatus"
      },
      "ReplayTarget": {
        "type": "string",
        "enum": [
          "kafka",
          "callback",
          "file",
          "test"
        ],
        "title": "ReplayTarget"
      },
      "ReplayType": {
        "type": "string",
        "enum": [
          "execution",
          "time_range",
          "event_type",
          "query",
          "recovery"
        ],
        "title": "ReplayType"
      },
...
      "SessionSummary": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "replay_type": {
            "$ref": "#/components/schemas/ReplayType"
          },
          "target": {
            "$ref": "#/components/schemas/ReplayTarget"
          },
          "status": {
            "$ref": "#/components/schemas/ReplayStatus"
          },
          "total_events": {
            "type": "integer",
            "title": "Total Events"
          },
          "replayed_events": {
            "type": "integer",
            "title": "Replayed Events"
          },
          "failed_events": {
            "type": "integer",
            "title": "Failed Events"
          },
          "skipped_events": {
            "type": "integer",
            "title": "Skipped Events"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds",
            "readOnly": true
          },
          "throughput_events_per_second": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Throughput Events Per Second",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "replay_type",
          "target",
          "status",
          "total_events",
          "replayed_events",
          "failed_events",
          "skipped_events",
          "created_at",
          "started_at",
          "completed_at",
          "duration_seconds",
          "throughput_events_per_second"
        ],
        "title": "SessionSummary",
        "description": "Summary information for replay sessions"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}